+++ test_opensearchpy/test_async/test_server/test_helpers/test_actions.py
@@ -60,6 +60,9 @@ class FailingBulkClient(object):
         self._fail_with = fail_with
 
     async def bulk(self, *args: Any, **kwargs: Any) -> Any:
+        """
+        increments number of times called and, when it equals fail_at, raises self.fail_with
+        """
         self._called += 1
         if self._called in self._fail_at:
             raise self._fail_with