+++ test_opensearchpy/test_connection/test_requests_http_connection.py
@@ -139,7 +139,8 @@ class TestRequestsHttpConnection(TestCase):
             )
             self.assertEqual(1, len(w))
             self.assertEqual(
-                "Connecting to https://localhost:9200 using SSL with verify_certs=False is insecure.",
+                "Connecting to https://localhost:9200 using SSL with "
+                "verify_certs=False is insecure.",
                 str(w[0].message),
             )
 
@@ -286,8 +287,9 @@ class TestRequestsHttpConnection(TestCase):
 
         # trace request
         self.assertEqual(1, tracer.info.call_count)
+        trace_curl_cmd = "curl -H 'Content-Type: application/json' -XGET 'http://localhost:9200/?pretty&param=42' -d '{\n  \"question\": \"what\\u0027s that?\"\n}'"  # pylint: disable=line-too-long
         self.assertEqual(
-            """curl -H 'Content-Type: application/json' -XGET 'http://localhost:9200/?pretty&param=42' -d '{\n  "question": "what\\u0027s that?"\n}'""",
+            trace_curl_cmd,
             tracer.info.call_args[0][0] % tracer.info.call_args[0][1:],
         )
         # trace response
@@ -415,9 +417,13 @@ class TestRequestsHttpConnection(TestCase):
         self.assertEqual('{"answer": 42}'.encode("utf-8"), request.body)
 
         # trace request
+        trace_curl_cmd = (
+            "curl -H 'Content-Type: application/json' -XGET 'http://localhost:9200/_search?pretty' "
+            "-d '{\n  \"answer\": 42\n}'"
+        )
         self.assertEqual(1, tracer.info.call_count)
         self.assertEqual(
-            "curl -H 'Content-Type: application/json' -XGET 'http://localhost:9200/_search?pretty' -d '{\n  \"answer\": 42\n}'",
+            trace_curl_cmd,
             tracer.info.call_args[0][0] % tracer.info.call_args[0][1:],
         )
 
@@ -514,7 +520,7 @@ class TestRequestsConnectionRedirect(TestCase):
 
     @classmethod
     def setup_class(cls) -> None:
-        # Start servers
+        """Start servers"""
         cls.server1 = TestHTTPServer(port=8080)
         cls.server1.start()
         cls.server2 = TestHTTPServer(port=8090)
@@ -522,7 +528,7 @@ class TestRequestsConnectionRedirect(TestCase):
 
     @classmethod
     def teardown_class(cls) -> None:
-        # Stop servers
+        """Stop servers"""
         cls.server2.stop()
         cls.server1.stop()