+++ utils/license_headers.py
@@ -47,11 +47,16 @@ def find_files_to_fix(sources: List[str]) -> Iterator[str]:
 
 
 def does_file_need_fix(filepath: str) -> bool:
+    """
+    checks if the correct license header exists at the top of the file
+    :param filepath: an absolute or relative filepath to a file to check
+    :return: True if the file needs a header, False if it does not
+    """
     if not re.search(r"\.py$", filepath):
         return False
     existing_header = ""
-    with open(filepath, mode="r") as f:
-        for line in f:
+    with open(filepath, mode="r", encoding="utf-8") as file:
+        for line in file:
             line = line.strip()
             if len(line) == 0 or line in LINES_TO_KEEP:
                 pass
@@ -64,20 +69,30 @@ def does_file_need_fix(filepath: str) -> bool:
 
 
 def add_header_to_file(filepath: str) -> None:
-    with open(filepath, mode="r") as f:
-        lines = list(f)
+    """
+    writes the license header to the beginning of a file
+    :param filepath: relative or absolute filepath to update
+    """
+    with open(filepath, mode="r", encoding="utf-8") as file:
+        lines = list(file)
     i = 0
     for i, line in enumerate(lines):
         if len(line) > 0 and line not in LINES_TO_KEEP:
             break
     lines = lines[:i] + [LICENSE_HEADER] + lines[i:]
-    with open(filepath, mode="w") as f:
-        f.truncate()
-        f.write("".join(lines))
+    with open(filepath, mode="w", encoding="utf-8") as file:
+        file.truncate()
+        file.write("".join(lines))
     print(f"Fixed {os.path.relpath(filepath, os.getcwd())}")
 
 
 def main() -> None:
+    """
+    arguments:
+        fix: find all files without license headers and insert headers at the top of the file
+        check: prints a list of files without license headers
+        list of one or more directories: search in these directories
+    """
     mode = sys.argv[1]
     assert mode in ("fix", "check")
     sources = [os.path.abspath(x) for x in sys.argv[2:]]